Add action method for Forms

+++ public_html/admin/list.php
@@ -3,6 +3,7 @@
 use App\App;
 use App\Views\BasePage;
 use App\Views\Forms\Admin\DeleteForm;
+use Core\Views\Form;
 use Core\Views\Link;
 use Core\Views\Table;
 
@@ -13,12 +14,14 @@ if (!App::$session->getUser()) {
     exit();
 }
 
-$deleteForm = new DeleteForm();
+if(Form::action()) {
+    $deleteForm = new DeleteForm();
 
-if ($deleteForm->validate()) {
-    $clean_inputs = $deleteForm->values();
+    if ($deleteForm->validate()) {
+        $clean_inputs = $deleteForm->values();
 
-    App::$db->deleteRow('items', $clean_inputs['id']);
+        App::$db->deleteRow('items', $clean_inputs['id']);
+    }
 }
 
 $rows = App::$db->getRowsWhere('items');
